waitlist-microservice/requirements.txt
Flask==3.0.0
Flask_JWT_Extended==4.5.3
pyodbc==4.0.39
python-dotenv==1.0.0